AB#600190 refactor: extract Jinja2 filters into dedicated template_filters module for improved organization
